requirements.txt
Django>=4.2.0
gunicorn>=20.1.0
python-decouple>=3.8
psycopg2-binary>=2.9.6
django-widget-tweaks>=1.5.0
django-filter>=23.5
django-summernote>=0.8.20.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7